Make sure we reject promise with an Error object
In the case of an error returned from O365 (non 200 status) it was rejected with the xhrResponse object which is not a proper Error object. I've made a suggested change to create an Error object with the status code in it.
We had issues with our service hanging with unresolved promised after a period of time.